requirements.txt
requests
pandas
schedule
python-telegram-bot==20.0